server part refactoring
1) fit on pep coding convention
2) renaming
